# src/swagger_documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DigitalStore API",
    "description": "Documentação da API da digital store"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "description": "JWT token",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/usuarios/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Autentica um usuário",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_email": {
                  "type": "string",
                  "example": "usuario@email.com"
                },
                "usuario_senha": {
                  "type": "string",
                  "example": "123456"
                }
              },
              "required": [
                "usuario_email",
                "usuario_senha"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login efetuado",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_nome": {
                  "type": "string",
                  "example": "chico moedas"
                },
                "usuario_email": {
                  "type": "string",
                  "example": "chico@gmail.com"
                },
                "token": {
                  "type": "string",
                  "example": "..."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Erro interno",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "detail": {
                  "type": "string",
                  "example": "mensagem do sistema"
                },
                "severity": {
                  "type": "string",
                  "example": "danger"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/usuarios/req-recuperar-senha": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/usuarios/recuperar-senha/{token}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_senha": {
                  "example": "any"
                },
                "usuarioEmail": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        }
      }
    },
...
  },
...
}